# board/internal/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "blerg-board",
    "version": "v1",
    "summary": "Kanban board where agents draft cards and humans curate the work.",
    "description": "The public, agent-facing contract of blerg-board (contract version v1). Paths are relative to this component's base URL, which GET /agents reports as base_url. These are the operations blerg-board's MCP tools wrap, so the same surface is reachable over plain HTTP or over MCP at /mcp with the same bearer token. The browser transport (WebSocket, diffs, metrics, icons), the operator surface (tokens, deployments, board deletion) and the session-driving routes are deliberately not described here: they are the product's own UI and control plane, not the card-writing contract.",
    "x-admission-gate": "Agent writes on a gated board pass an admission gate: 409 with a review_id when denied, 202 with a review_id when held for a human. Read the body and act on it — resubmit once with dispute_of and a rebuttal to contest a denial."
  },
  "servers": [],
...
  "paths": {
...
    "/api/boards": {
      "get": {
        "operationId": "list_boards",
        "tags": ["boards"],
        "summary": "List the boards this credential can see.",
        "description": "A board-scoped token sees only its own board.",
        "security": [{ "agentToken": ["card.read"] }, { "boardToken": ["card.read"] }],
        "responses": {
          "200": {
            "description": "The visible boards.",
            "content": {
              "application/json": { "schema": { "type": "array", "items": { "$ref": "#/components/schemas/Board" } } }
            }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" }
        }
      },
      "post": {
        "operationId": "create_board",
        "tags": ["boards"],
        "summary": "Create a board.",
        "description": "Instance-wide admin work: requires `board.admin`.",
        "security": [{ "agentToken": ["board.admin"] }, { "boardToken": ["board.admin"] }],
        "requestBody": {
          "required": true,
          "content": { "application/json": { "schema": { "$ref": "#/components/schemas/BoardInput" } } }
        },
        "responses": {
          "201": {
            "description": "The created board.",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Board" } } }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "422": { "$ref": "#/components/responses/Unprocessable" }
        }
      }
    },
    "/api/boards/{id}": {
      "get": {
        "operationId": "get_board",
        "tags": ["boards"],
        "summary": "One board: name, repos, gate policy, columns and field schema.",
        "security": [{ "agentToken": ["card.read"] }, { "boardToken": ["card.read"] }],
        "parameters": [{ "$ref": "#/components/parameters/BoardId" }],
        "responses": {
          "200": {
            "description": "The board.",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Board" } } }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "404": { "$ref": "#/components/responses/NotFound" }
        }
      },
      "patch": {
        "operationId": "update_board",
        "tags": ["boards"],
        "summary": "Update board settings, repos, models or field schema.",
        "security": [{ "agentToken": ["board.admin"] }, { "boardToken": ["board.admin"] }],
        "parameters": [{ "$ref": "#/components/parameters/BoardId" }],
        "requestBody": {
          "required": true,
          "content": { "application/json": { "schema": { "$ref": "#/components/schemas/BoardInput" } } }
        },
        "responses": {
          "200": {
            "description": "The updated board.",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Board" } } }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "404": { "$ref": "#/components/responses/NotFound" },
          "422": { "$ref": "#/components/responses/Unprocessable" }
        }
      }
    },
...
  },
  "components": {
    "parameters": {
      "BoardId": {
        "name": "id",
        "in": "path",
        "required": true,
        "description": "The board's id.",
        "schema": { "type": "string", "format": "uuid" }
      },
...
    },
    "securitySchemes": {
      "agentToken": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "A core-signed token with `aud: \"blerg-board\"`, minted from the `board` preset. Get one from core: POST /api/tokens, or Settings → Agent tokens. A human's core-issued session token authenticates the same way and is capability-checked identically."
      },
      "boardToken": {
        "type": "http",
        "scheme": "bearer",
        "description": "A board-native token (`blergb_…`), minted by an admin or handed to a session when it is spawned. May also be sent as the X-Blerg-Board-Token header. Carries an explicit board scope and capability set."
      }
    },
    "responses": {
...
      "BadRequest": { "description": "Malformed request body.", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } },
      "Unauthorized": { "description": "Missing, malformed, expired or revoked credential.", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } },
      "Forbidden": { "description": "The credential is valid but lacks the capability, or is scoped to another board.", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } },
      "NotFound": { "description": "No such board, column, card or review.", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } },
      "Unprocessable": { "description": "The request was understood but rejected: an unknown field key, a type mismatch against the board's schema, or a repo the board does not declare.", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } },
...
    },
    "schemas": {
      "Error": {
        "type": "object",
        "required": ["error"],
        "properties": {
          "error": { "type": "string" },
          "field": { "type": "string", "description": "The offending field key, on a schema rejection." }
        }
      },
...
      "Board": {
        "type": "object",
        "required": ["id", "name"],
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "name": { "type": "string" },
          "description": { "type": ["string", "null"] },
          "repos": { "type": "array", "items": { "type": "string" } },
          "gate_enabled": { "type": "boolean", "description": "Whether agent writes pass the admission gate." },
          "require_repo": { "type": "boolean" },
          "field_schema": { "type": "array", "items": { "$ref": "#/components/schemas/FieldDef" } },
          "concurrency": { "type": "integer", "description": "How many cards a board run works at once." },
          "deploy_url": { "type": ["string", "null"] },
          "automation_engine": { "type": "string", "enum": ["claude", "codex", "hermes"], "description": "Engine every session this board starts runs." },
          "automation_token_set": { "type": "boolean", "description": "Whether an automation token is configured. The token itself is write-only and never returned." },
          "automation_token_expires_at": { "type": ["string", "null"], "format": "date-time", "description": "When the configured automation token expires." },
          "automation_account_id": { "type": ["string", "null"], "description": "The blerg-core account the automation token acts for: whose engine credential every session this board starts spends." }
        }
      },
      "BoardInput": {
        "type": "object",
        "properties": {
          "name": { "type": "string" },
          "description": { "type": "string" },
          "repos": { "type": "array", "items": { "type": "string" } },
          "columns": { "type": "array", "items": { "type": "string" }, "description": "Initial column names, in order (create only)." },
          "field_schema": { "type": "array", "items": { "$ref": "#/components/schemas/FieldDef" } },
          "gate_enabled": { "type": "boolean" },
          "require_repo": { "type": "boolean" },
          "concurrency": { "type": "integer" },
          "deploy_url": { "type": "string" },
          "automation_engine": { "type": "string", "enum": ["claude", "codex", "hermes"] },
          "automation_token": { "type": "string", "writeOnly": true, "description": "PATCH only, and only by a signed-in person: a blerg-core agent token (preset run-sessions) they minted for themselves. Every session the board starts runs as that person, on their own engine credential. Empty string clears it. Never returned." }
        }
      },
      "FieldDef": {
        "type": "object",
        "required": ["key", "type"],
        "properties": {
          "key": { "type": "string" },
          "label": { "type": "string" },
          "type": { "type": "string", "enum": ["enum", "text", "number", "url", "timestamp", "bool"] },
          "values": { "type": "array", "items": { "type": "string" } },
          "display": { "type": "string", "enum": ["badge", "chip", "link", "inline", "hidden"] },
          "filterable": { "type": "boolean" },
          "order": { "type": "integer" }
        }
      },
...
    }
  }
}